pyproject.toml
[tool.poetry]
name = "marktplaats-messages"
version = "0.3.0"
description = "Marktplaats Messages unofficial client"
authors = [
  "Aleksandr Vinokurov <aleksandr.vin@gmail.com>"
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/aleksandr-vin/marktplaats-messages"


[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/aleksandr-vin/marktplaats-messages/issues"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-ra --cov=marktplaats_messages"
testpaths = [
    "tests",
]